Update the sample's version and structure to be closer to the kmp.jetbrains.com wizard's structure.
This is not the final state of the refactorings. I still want to add a desktop version, which will use the CMP UI. I also want to change the patterns we're using to get the data to more MVVM aligned. At the moment iOS specifically is not this and some users were complaining.